Add files via upload
A project that showcased the power of Python in data cleaning and analysis. The objective of this project was to transform a complex dataset into actionable insights for business stakeholders. In this report, I will provide an overview of the project, discuss the technical details, present a few visualizations, results, and conclude with the impact of the project. Additionally, I will provide recommendations for the business based on the findings. 
Link to the dataset: https://github.com/AlexTheAnalyst/PandasYouTubeSeries/blob/main/Customer%20Call%20List.xlsx